+++ src/main/java/webserver/message/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.message;
 
 import java.util.Map;
