+++ src/main/java/webserver/message/URI.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.message;
 
 import java.util.HashMap;
 import java.util.List;
@@ -33,5 +33,17 @@ public class URI {
 
     public boolean hasParameter() {
         return !parameters.isEmpty();
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    @Override
+    public String toString() {
+        return "URI{" +
+                "path='" + path + '\'' +
+                ", parameters=" + parameters +
+                '}';
     }
 }
